requirements.txt
Jinja2==3.1.4
PyYAML==6.0.2
jsonschema==4.23.0
typer==0.12.3
rich==13.8.0
tomli==2.0.1
pytest==8.3.2
requests==2.32.3
orjson==3.10.7
packaging==24.1
colorama==0.4.6
dataclasses-json==0.6.7
